adjustments

+++ src/main/java/com/example/market/service/ProductService.java
@@ -1,6 +1,7 @@
 package com.example.market.service;
 
 import com.example.market.controller.exceptionHandler.exceptions.ClientNotFoundException;
+import com.example.market.controller.exceptionHandler.exceptions.ProductHasDifferentOwnerException;
 import com.example.market.controller.exceptionHandler.exceptions.ProductNotFoundException;
 import com.example.market.entity.model.Product;
 import com.example.market.repository.ClientRepository;
@@ -9,6 +10,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -38,12 +40,24 @@ public class ProductService {
     public Product findProductById(Long clientId, Long productId) {
         clientService.checkIfClientExists(clientRepository.findById(clientId), clientId);
         Optional<Product> optionalProduct = productRepository.findById(productId);
-        return optionalProduct.orElseThrow(() -> new ProductNotFoundException(Constant.PRODUCT_NOT_FOUND_ERROR_MESSAGE + productId));
+        checkProductPresenceAndOwner(optionalProduct, productId, clientId);
+        return optionalProduct.orElse(null);
     }
 
     @Transactional
     public void deleteOne(Long clientId, Long productId) {
         clientService.checkIfClientExists(clientRepository.findById(clientId), clientId);
+        checkProductPresenceAndOwner(productRepository.findById(productId), productId, clientId);
         productRepository.deleteById(productId);
+    }
+
+    // Проверка на существование товара и его принадлежность к данному клиенту
+    public void checkProductPresenceAndOwner(Optional<Product> optionalProduct, Long productId, Long clientId) {
+        if (optionalProduct.isEmpty()) {
+            throw new ProductNotFoundException(Constant.PRODUCT_NOT_FOUND_ERROR_MESSAGE + productId);
+        }
+        if (!Objects.equals(optionalProduct.get().getClient().getId(), clientId)) {
+            throw new ProductHasDifferentOwnerException(Constant.PRODUCT_HAS_DIFFERENT_OWNER_ERROR_MESSAGE);
+        }
     }
 }
